refactor: move test to co-folder
