+++ src/commands/moderators/stickyMessage.spec.ts
@@ -11,13 +11,14 @@ import {
 import { StickyMessage } from '@prisma/client'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 
-import stickyMessage from '../../../../src/commands/moderators/stickyMessage.js'
-import { STICKY_CACHE_PREFIX } from '../../../../src/features/stickyMessage/index.js'
-import { prisma } from '../../../../src/prisma.js'
-import { BotContext } from '../../../../src/types/BotContext.js'
-import * as cache from '../../../../src/utils/cache.js'
+import { STICKY_CACHE_PREFIX } from '../../features/stickyMessage/index.js'
+import { prisma } from '../../prisma.js'
+import { BotContext } from '../../types/BotContext.js'
+import * as cache from '../../utils/cache.js'
 
-vi.mock('../../../../src/config.js', async () => {
+import stickyMessage from './stickyMessage.js'
+
+vi.mock('../../config.js', async () => {
   const Environment = {}
 
   return { Environment }
