+++ src/events/stickyMessageRemove.spec.ts
@@ -5,11 +5,12 @@ import { prisma } from '@/prisma.js'
 import { StickyMessage } from '@prisma/client'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 
-import stickyMessage from '../../../src/events/stickyMessageRemove.js'
-import { BotContext } from '../../../src/types/BotContext.js'
-import * as cache from '../../../src/utils/cache.js'
+import { BotContext } from '../types/BotContext.js'
+import * as cache from '../utils/cache.js'
 
-vi.mock('../../../src/config.js', async () => {
+import stickyMessage from './stickyMessageRemove.js'
+
+vi.mock('../config.js', async () => {
   const Environment = {
     MESSAGE_COOLDOWN_SEC: 15,
     MESSAGE_MAX: 5,
