+++ src/events/stickyMessageSet.spec.ts
@@ -3,13 +3,14 @@ import { ChannelType, Client, Message, TextChannel } from 'discord.js'
 import { StickyMessage } from '@prisma/client'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 
-import stickyMessage from '../../../src/events/stickyMessageSet.js'
-import { STICKY_CACHE_PREFIX } from '../../../src/features/stickyMessage/index.js'
-import { prisma } from '../../../src/prisma.js'
-import { BotContext } from '../../../src/types/BotContext.js'
-import * as cache from '../../../src/utils/cache.js'
+import { STICKY_CACHE_PREFIX } from '../features/stickyMessage/index.js'
+import { prisma } from '../prisma.js'
+import { BotContext } from '../types/BotContext.js'
+import * as cache from '../utils/cache.js'
 
-vi.mock('../../../src/config.js', async () => {
+import stickyMessage from './stickyMessageSet.js'
+
+vi.mock('../config.js', async () => {
   const Environment = {}
 
   return { Environment }
